PER-8805: Update PR requested changes

+++ packages/api/src/account/queries/delete_account_archive.sql
@@ -1,6 +1,7 @@
 DELETE FROM account_archive
 WHERE
   archiveid = :archiveId
+  AND accessrole != 'access.role.owner'
   AND accountid IN (SELECT accountid FROM account WHERE primaryemail = :email)
 RETURNING
 account_archiveid AS "accountArchiveId";
